Tentativa de implementar lógica de bloco de escopo no EguaP.
